// src/lib/embody/lib/validating/types.ts
type SourcePosition = {
	readonly line: number;
	readonly column: number;
};
// ...
type SourceRange = {
	readonly start: SourcePosition;
	readonly end: SourcePosition;
};
// ...
type Violation = {
	readonly nodeType: string;
	readonly message: string;
	readonly severity: 'rejection';
	readonly location: SourceRange;
	readonly nodePath: string;
};
// ...
type NodeValidator = (node: Node, nodePath: string) => true | Violation;
// ...
type NodeRule = true | false | NodeValidator;
// ...
type SyntaxAllowlist = {
	readonly name: string;
	readonly allowedGlobals?: ReadonlySet<string>;
	readonly blockedMemberNames?: ReadonlySet<string>;
	readonly nodes: Readonly<Record<string, NodeRule>>;
};